organize random files
