+++ services/ui-src/src/components/wrappers/App.tsx
@@ -1,4 +1,4 @@
-import { AppRoutes } from "./Routes";
+import AppRoutes from "./AppRoutes";
 import * as QMR from "components";
 import { LocalLogins } from "components";
 import { useUser } from "hooks/authHooks";
