+++ services/ui-src/src/components/wrappers/QueryProvider.tsx
@@ -2,8 +2,10 @@ import { QueryClient, QueryClientProvider } from "react-query";
 
 const queryClient = new QueryClient();
 
-export const QueryProvider = ({ children }: any) => {
+const QueryProvider = ({ children }: any) => {
   return (
     <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
   );
 };
+
+export default QueryProvider;
