--- services/ui-src/src/setupTests.ts
@@ -1,7 +1,0 @@
-/*
- * jest - dom adds custom jest matchers for asserting on DOM nodes.
- * allows you to do things like:
- * expect(element).toHaveTextContent(/react/i)
- * learn more: https://github.com/testing-library/jest-dom
- */
-import "@testing-library/jest-dom/extend-expect";
