+++ services/ui-src/src/utils/serviceWorker.ts
